refactor: create utils folder for common functions and config parametrs
