+++ utils/config/open_ai.py
@@ -1,41 +1,6 @@
-
-# config.py
-
-JIRA_CONFIG = {
-    "server":  "",
-    "username": "",
-    "token": ""
-}
-
-OPENAI_CONFIG = {
+OPENAI = {
     "api_key": "",
     "model": "gpt-4o"
-}
-
-PARAMETERS = {
-    "project": "",
-    "issuetype": "",
-    "sprint": ""
-}
-
-QUERIES = {
-    "ticket_scores": "project = {project} AND issuetype = {issuetype} AND Sprint = {sprint}",
-    "backlog": "project = {project} AND Sprint IS EMPTY AND statusCategory != Done"
-}
-
-TEMPLATE_HEADINGS = {
-    "task_template_headings": [
-        "Summary",
-        "Acceptability Criteria",
-        "Technical Notes",
-        "Testing Instructions"
-    ],
-    "bug_template_headings": [
-        "Summary",
-        "Steps to Reproduce",
-        "Technical Notes",
-        "Testing Instructions"
-    ]
 }
 
 PROMPTS = {
@@ -65,24 +30,4 @@ PROMPTS = {
         Please restructure and rephrase the description to fit perfectly under each heading with 100 percent adherence.
         Provide the revised description with sections clearly separated by headings.
     '''
-}
-
-SCORING = {
-    "weightage_of_relevance": 0.5,
-    "weightage_of_adherence": 0.5
-}
-
-TEMPLATE_PLACEHOLDERS = {
-    "bug": {
-        "summary": "[A concise description of the issue or task.]",
-        "steps_to_reproduce": "[Detailed steps that lead to the issue, including screenshots if possible.]",
-        "technical_notes": "[Any technical information that might be helpful for the developer, such as dependencies, libraries, or code snippets.]",
-        "testing_instructions": "[Instructions or considerations for testing the task or fix.]"
-    },
-    "task": {
-        "summary": "[A concise description of the task.]",
-        "steps_to_reproduce": "[Detailed steps to complete the task.]",
-        "technical_notes": "[Technical information relevant to the task.]",
-        "testing_instructions": "[Instructions for verifying the task's completion.]"
-    }
 }
